% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{allClassMeansToGR}
\alias{allClassMeansToGR}
\title{Convert allClassMeans table to GenomicRanges}
\usage{
allClassMeansToGR(allClassMeans, regionGR, anchorPoint = "middle", dSMFwin = 1)
}
\arguments{
\item{allClassMeans}{A long data frame of methylation or bincount values with columns for position, methylation frequency (methFreq), class and replicate.}

\item{regionGR}{A genomic range for the region that is being plotted}

\item{anchorPoint}{One of "middle" or "start": the position from which numbering starts}

\item{dSMFwin}{Width of dSMF score window (default is 1bp)}
}
\value{
Returns a GenomicRanges object with absolute genomic coordinates and
methylation score for different classes and replicates in mcols
}
\description{
Convert allClassMeans table to GenomicRanges with absolute genomic
coordinates and with methylation score at each position for each class and
multiple replicates in the metadata columns (required for plotting with
Gviz)
}
