% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{classifyAndSortReads}
\alias{classifyAndSortReads}
\title{Classify and sort reads into their classes}
\usage{
classifyAndSortReads(dataMatrix, posteriorProb, previousClassMeans = NULL)
}
\arguments{
\item{dataMatrix}{A matrix of methylation or bincount values (reads x position)}

\item{posteriorProb}{posteriorProb: a matrix of probabilites of each sample belonging to a particular class (samples x class)}

\item{previousClassMeans}{A matrix of the class means from a previous round of clustering}
}
\value{
Returns a matrix with the reads classified (__classX is appended to the read name), and the classes are sorted.
}
\description{
Classify reads by their posterior probability of belonging to a specific class
Then sort the classes by using similarity to the mean profile of previous classes.
If the means of previous classes was not provided, hclust is used to cluste classes by their similarity.
}
