% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyByKnownClasses.R
\name{classLikelihoodPerRead}
\alias{classLikelihoodPerRead}
\title{Get likelihood of different classes for each read}
\usage{
classLikelihoodPerRead(dataMatrix, classes)
}
\arguments{
\item{dataMatrix}{A matrix of methylation or bincount values (reads x position)}

\item{classes}{A matrix of classes x position giving the mean frequency of methylation at each position for each class}
}
\value{
likelihood of different classes for each read
}
\description{
Get likelihood of different classes for each read
}
