% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyByKnownClasses.R
\name{combineSilStats}
\alias{combineSilStats}
\title{Combine read counts in each class from silStats files}
\usage{
combineSilStats(matTable, numClasses, outPath)
}
\arguments{
\item{matTable}{Table with list of all samples and regions which have been
classified according to known classes.}

\item{numClasses}{An integer indicating the total number of classes being used}

\item{outPath}{A string with the path to the directory where the output should go}
}
\value{
data.frame of the number of reads in each class, with extra columns for
total number of reads, region and sample names
}
\description{
Function takes all region_sample listed in matTable and extracts the read
counts in each class from the single line silStats files that are generated
when clustering according to known classes.
}
