% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{correlationDist}
\alias{correlationDist}
\title{Calculate correlation distance between all rows of a matrix}
\usage{
correlationDist(binMat, valNA = NULL, rescale = F)
}
\arguments{
\item{binMat}{A matrix of numbers for which you want to calculate the
distance between rows}

\item{valNA}{Value to give NAs in matrix}

\item{rescale}{Should matrix be rescaled from 0..1 range to -1..1 range?}
}
\value{
A distance object (lower triangle) with the distances between all
rows of the input matrix
}
\description{
Using the correlation function from stats package
}
