% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{crossCorDist}
\alias{crossCorDist}
\title{Calculate cross correlation distance between all rows of a matrix}
\usage{
crossCorDist(binMat)
}
\arguments{
\item{binMat}{A matrix of numbers for which you want to calculate the
distance between rows}
}
\value{
A distance object (lower triangle) with the distances between all
rows of the input matrix
}
\description{
Using the cosine distance function form las package
}
