% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{em_basic}
\alias{em_basic}
\title{Expectation Maximisation algorithm}
\usage{
em_basic(classes, priorProb, dataMatrix)
}
\arguments{
\item{classes}{A matrix containing the classes to be optimised. c[i,j] is the expected methylation value or bincount value of class i at position j.}

\item{priorProb}{A vector defining the prior probabilities of each class.}

\item{dataMatrix}{A matrix containing the sample. dataMatrix [i,j] is the methylation value or bincount value of sample i at position j.}
}
\value{
A list of three items:
            1) classes: a matrix with classes being optimised (class x position)
            2) priorProb: a vector with the prior probabilities of each class
            3) posteriorProb: a matrix of probabilites of each sample belonging
            to a particular class (samples x class)
}
\description{
Expectation Maximisation algorithm
}
