% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{euclidean}
\alias{euclidean}
\title{Calculate Euclidean distance between two vectors}
\usage{
euclidean(r1, r2)
}
\arguments{
\item{r1}{A numeric vector}

\item{r2}{Second numeric vector of same length as r1}
}
\value{
Euclidean distance between two vectors
}
\description{
Calculate Euclidean distance between two vectors
}
