% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{euclidWin}
\alias{euclidWin}
\title{Calculate Euclidean distance between two vectors with sliding windows}
\usage{
euclidWin(r1, r2, winSize = 3)
}
\arguments{
\item{r1}{A numeric vector}

\item{r2}{Second numeric vector of same length as r1}

\item{winSize}{Sliding window size (number of values to combine)}
}
\value{
Euclidean distance between two vectors
}
\description{
To increase dynamic range and spatial information in comparin two binary
vectors, a sliding window is used. The euclidean distance is computed within
each window, and then moved by a step of 1.
}
