% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{euclidWinDist}
\alias{euclidWinDist}
\title{Calculate Euclidean distance between all rows of a matrix with sliding window}
\usage{
euclidWinDist(binMat, winSize = 3, stepSize = 1)
}
\arguments{
\item{binMat}{A matrix of numbers for which you want to calculate the
distance between rows}

\item{winSize}{Sliding window size (number of values to combine)}

\item{stepSize}{How far to move the sliding window}
}
\value{
A distance object (lower triangle) with the distances between all
rows of the input matrix
}
\description{
To increase dynamic range and spatial information in comparing a matrix of
vectors, a sliding window is used. The euclidean distance is computed within
each window, and then moved by a step of 1.
}
