% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyByKnownClasses.R
\name{getAllClassMeansMatrix}
\alias{getAllClassMeansMatrix}
\title{Collect all the class means from allClassMeans object}
\usage{
getAllClassMeansMatrix(
  allClassMeans,
  k_range = 4:8,
  xRange = c(-250, 250),
  name = ""
)
}
\arguments{
\item{allClassMeans}{A list by number of classes of tables containing positions, fraction methylation, and class ID for replicate EM runs.}

\item{k_range}{An vector of the range of total number of classes used in each clustering}

\item{xRange}{A vector of the first and last coordinates of the region to plot (default is c(-250,250))}

\item{name}{Name of data set that will be added to the start of the row names.}
}
\value{
A matrix of all the class mean profiles from the clustering into all the k clusters in k_range
}
\description{
AllClassMeans is a list of tables with the position in the list corresponding to the total number of classes used in clustering and the table giving the methylation frequency at different positions for each class and each replicate. This function collects all the class mean profiles for clustering into any k clusters in k_range, and returns a matrix with cluster-name-number x position
}
