% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{getClasses}
\alias{getClasses}
\title{Extract class info from dataOrderedByClass}
\usage{
getClasses(dataOrderedByClass, readNames)
}
\arguments{
\item{dataOrderedByClass}{A matrix of methylation frequency or bin counts for
indivudal reads at particular positions where the reads have been sorted by class and
the row names contain the read name and the class joined together: readName__classX}

\item{readNames}{A vector of read names by which to order the classes}
}
\value{
Classification of reads ordered by readNames
}
\description{
Extract class info from dataOrderedByClass
}
