% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyByKnownClasses.R
\name{getClassMeansMatrix}
\alias{getClassMeansMatrix}
\title{Get a matrix of class mean profiles from allCalssMeans table}
\usage{
getClassMeansMatrix(allClassMeans, numClasses)
}
\arguments{
\item{allClassMeans}{A list by number of classes of tables containing positions, fraction methylation, and class ID for replicate EM runs.}

\item{numClasses}{An integer indicating the total number of classes being used}
}
\value{
A matrix with classID x position containing fraction methylation values for each class at each position
}
\description{
allClassMeans is a list of tables (one table for each number of classes). The function extracts one such table and converts it to a matrix with each row being a particular class and each column being a genomic position (in coordinates relative to an anchor point, like a TSS).
}
