% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{getClassVote}
\alias{getClassVote}
\title{Get most frequently assigned class from repeat runs}
\usage{
getClassVote(classVote)
}
\arguments{
\item{classVote}{A data.frame of reads with the class assignments from multiple EM repeats}
}
\value{
A data.frame including the most frequently called class for each read
}
\description{
Get most frequently assigned class from repeat runs
}
