% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmultigene.R
\name{getFullMatrix}
\alias{getFullMatrix}
\title{Get full matrix of all positions in a window even if no C}
\usage{
getFullMatrix(dataMatrix, winSize = 500, anchorPoint = "middle")
}
\arguments{
\item{dataMatrix}{A matrix of methylation values (reads x position) with
positions being relative coordinates to an anchor point}

\item{winSize}{The size (in bp) of the window containing the matrices}

\item{anchorPoint}{One of "middle" or "start": the position from which numbering starts for the positions in the dataMatrix}
}
\value{
A padded methylation matrix
}
\description{
In order to compare different promoters, we need to create a padded
matrix with NAs in positions in between Cs.
}
