% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMode}
\alias{getMode}
\title{Get mode (most frequent value)}
\usage{
getMode(v)
}
\arguments{
\item{v}{vector of values}
}
\value{
Value that appears most frequently
}
\description{
Function copied from https://www.tutorialspoint.com/r/r_mean_median_mode
}
