% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{getSilhouetteStats}
\alias{getSilhouetteStats}
\title{Draw silhouette plots and get silhouette statistics}
\usage{
getSilhouetteStats(
  dataOrderedByClass,
  numClasses,
  outFileBase,
  outPath,
  EMrep = NULL,
  doIndividualPlots = FALSE,
  silStats = NULL,
  distMetric = list(name = "euclidean", rescale = F)
)
}
\arguments{
\item{dataOrderedByClass}{A matrix of methylation or bincount values (reads x position) that have been ordered by class. The assigned class, e.g. "__class1" etc has been appended to read names.}

\item{numClasses}{An integer indicating the number of classes to learn}

\item{outFileBase}{A string that will be used in the filenames and titles of the plots produced}

\item{outPath}{path to directory where plots will be saved}

\item{EMrep}{An integer indicating which EM repeat this is}

\item{doIndividualPlots}{Produce individual plots for each repeat (default=F)}

\item{silStats}{data.frame with statistics about the silhouette plots (default=NULL)}

\item{distMetric}{A list with the name of the distance metric and any
parameters it might require}
}
\value{
silStats data.frame with statistics about the silhouette plots
}
\description{
Draw silhouette plots and get silhouette statistics
}
