% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{isMatrixValid}
\alias{isMatrixValid}
\title{Test if matrix is valid}
\usage{
isMatrixValid(dataMatrix, valueRange = c(0, 1), NAsValid = FALSE)
}
\arguments{
\item{dataMatrix}{Matrix the values of which will be tested}

\item{valueRange}{Vector of min and max permissible values in the matrix
(default=c(0,1))}

\item{NAsValid}{Boolean TRUE/FALSE indicating if NAs are considered valid (default=FALSE)}
}
\value{
Boolean TRUE/FALSE indicating if matrix contains only valid values
}
\description{
Test if matrix contains values that are only within the range specified by valueRange
}
