% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{order_by_prev_cluster}
\alias{order_by_prev_cluster}
\title{Order classes}
\usage{
order_by_prev_cluster(numClasses, classMeans, prev_classMeans)
}
\arguments{
\item{numClasses}{An integer with the number of classes learned}

\item{classMeans}{The mean profile of all reads in their classes}

\item{prev_classMeans}{The mean profile of all reads in their classes from a previous run}
}
\value{
Returns a vector with the order of the classes
}
\description{
Order classes either according to the order of previous classes or order by similarity using hclust
}
