% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{p_converged}
\alias{p_converged}
\title{Check if p value has converged}
\usage{
p_converged(p_diff, p_diff_prev, error)
}
\arguments{
\item{p_diff}{The difference in likelihood betwween this round of optimisation and the previous round.}

\item{p_diff_prev}{The difference in likelihood between the previous round of optimisation and the one before that.}

\item{error}{The maximum tolerated error}
}
\value{
A TRUE or FALSE value indicating if the difference between successive rounds of optimisation is below the error threshold (i.e. converged)
}
\description{
Check if p value has converged
}
