% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{p_difference}
\alias{p_difference}
\title{Get absolute value of difference between two p values}
\usage{
p_difference(p, p_prev)
}
\arguments{
\item{p}{Current probability}

\item{p_prev}{Probability from last round of optimisation}
}
\value{
The absolute value of the change in probability
}
\description{
Get absolute value of difference between two p values
}
