% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessMethMats.R
\name{padMissingColWithNAs}
\alias{padMissingColWithNAs}
\title{Add padding between informative positions.}
\usage{
padMissingColWithNAs(mat, colRange = c(-250, 250))
}
\arguments{
\item{mat}{A matrix of numbers and NAs.}

\item{colRange}{a vector of two numbers for start and end positions used in the matrix (default=c(-250,250))}
}
\value{
A matrix with missing columns padded with NAs
}
\description{
In order to combine matrices from different genes, the simple
read x position matrix must be converted to a full matrix with
a column for every position within the region, even if no CpG or GpC
is found there in a particular gene. Any non-informative position gets
the value "NA".
}
\examples{
m<-matrix(c(1:4),nrow=2,byrow=TRUE,dimnames=list(c("a","b"),c(-1,2)))
padMissingColWithNAs(m,colRange=c(-2,2))
}
