% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{plotClassesSingleGene}
\alias{plotClassesSingleGene}
\title{Plot reads by class for a single gene}
\usage{
plotClassesSingleGene(
  dataOrderedByClass,
  xRange = c(-250, 250),
  title = "Reads by classes",
  myXlab = "CpG/GpC position",
  featureLabel = "TSS",
  baseFontSize = 12
)
}
\arguments{
\item{dataOrderedByClass}{A matrix of methylation or bincount values (reads x position) that have been ordered by class. The assigned class, e.g. "__class1" etc has been appended to read names.}

\item{xRange}{A vector of the first and last coordinates of the region to plot (default is c(-250,250))}

\item{title}{A title for the plot (default is "Reads by classes")}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS (default="TSS")}

\item{baseFontSize}{The base font for the plotting theme (default=12 works well for 4x plots per A4 page)}
}
\value{
Returns a ggplot2 object of a single molecule plot sorted by classes
}
\description{
Create a single molecule plot of the reads sorted by class.
}
