% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{plotClassMeans}
\alias{plotClassMeans}
\title{Plot class Means}
\usage{
plotClassMeans(
  classes,
  xRange = c(-250, 250),
  facet = TRUE,
  title = "Class means",
  myXlab = "CpG/GpC position",
  featureLabel = "TSS",
  baseFontSize = 12
)
}
\arguments{
\item{classes}{A matrix of methylation or bincount values (classes x position) for each class}

\item{xRange}{A vector of the first and last coordinates of the region to plot (default is c(-250,250))}

\item{facet}{Plot mean profiles separately as a facet_wrap plot (default=TRUE)}

\item{title}{A title for the plot (default is "Class means")}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS (default="TSS")}

\item{baseFontSize}{The base font for the plotting theme (default=12 works well for 4x plots per A4 page)}
}
\value{
Returns a ggplot2 object
}
\description{
Create line plots for class means
}
