% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{plotClassMeansWithTracks}
\alias{plotClassMeansWithTracks}
\title{Plot smoothed class means from multiple repeats}
\usage{
plotClassMeansWithTracks(
  allClassMeans,
  regionGR,
  anchorPoint = "middle",
  txdb = NULL,
  featureGR = NULL,
  bigwigListFile = NULL,
  strandedBwFile = NULL
)
}
\arguments{
\item{allClassMeans}{A long data frame of methylation or bincount values with columns for position, methylation frequency (methFreq), class and replicate.}

\item{regionGR}{A genomic range for the region that is being plotted}

\item{anchorPoint}{One of "middle" or "start": the position from which numbering starts for the position in allClassMatrix}

\item{txdb}{A TxDb object e.g. TxDb.Celegans.UCSC.ce11.refGene from which to get the gene regions}

\item{featureGR}{A genoimc range for a feature you wish to highlight (e.g. TSS). Will be overlaid on transcript track.}

\item{bigwigListFile}{A chacter string with the name of a file that contains a list of bigwigs to be plotted with the class means. First column of this file must be the file name of the bigwig with the full path. the second column is the dataset name (will be used on the plot so keep it short).}

\item{strandedBwFile}{A chacter string with the name of a file that contains a list of stranded bigwigs to be plotted with the class means. First column of this file must be the file name of the bigwig with the full path. The second column is the dataset name (will be used on the plot so keep it short), third column is "+" or "-" indicating the strand of the bigwig.}
}
\value{
Returns a ggplot2 object
}
\description{
Plot loess-smoothed class means from multiple repeats of EM.
}
