% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{plotClassStability}
\alias{plotClassStability}
\title{Plot histograms of consistency of classfiications over multiple repeats}
\usage{
plotClassStability(classVote, outFileBase, outPath, numClasses)
}
\arguments{
\item{classVote}{A data.frame of reads with the class assignments from multiple EM repeats}

\item{outFileBase}{A string that will be used in the filenames and titles of the plots produced}

\item{outPath}{path to directory where plots will be saved}

\item{numClasses}{An integer indicating the number of classes to learn}
}
\value{
classVote data frame with "topClass" and "topClassFreq" columns added
}
\description{
Plot histograms of consistency of classfiications over multiple repeats
}
