% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{plotEachRepeat}
\alias{plotEachRepeat}
\title{Plot single molecule and class means for individual repeats}
\usage{
plotEachRepeat(
  dataOrderedByClass,
  outFileBase,
  outPath,
  numClasses,
  EMrep,
  classMeans,
  xRange,
  myXlab,
  featureLabel,
  baseFontSize
)
}
\arguments{
\item{dataOrderedByClass}{A matrix of methylation or bincount values (reads x position) that have been ordered by class. The assigned class, e.g. "__class1" etc has been appended to read names.}

\item{outFileBase}{A string that will be used in the filenames and titles of the plots produced}

\item{outPath}{path to directory where plots will be saved}

\item{numClasses}{An integer indicating the number of classes to learn}

\item{EMrep}{An integer indicating which repeat of the EM this is}

\item{classMeans}{A matrix (classes x position) of the average methlyation profile of each class classes}

\item{xRange}{A vector of the first and last coordinates of the region to plot (default is c(-250,250))}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS (default="TSS")}

\item{baseFontSize}{The base font for the plotting theme (default=12 works well for 4x plots per A4 page)}
}
\value{
Returns a ggplot2 object
}
\description{
Plot single molecule and class means for individual repeats
}
