% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{plotFinalClasses}
\alias{plotFinalClasses}
\title{Plot reads and class means with final classification}
\usage{
plotFinalClasses(
  dataOrderedByClass,
  numClasses,
  allClassMeans,
  outFileBase,
  outPath,
  xRange,
  myXlab,
  featureLabel,
  baseFontSize,
  figFormat = "png",
  classesToPlot = NULL
)
}
\arguments{
\item{dataOrderedByClass}{A matrix of methylation or bincount values (reads x position) that have been ordered by class. The assigned class, e.g. "__class1" etc has been appended to read names.}

\item{numClasses}{An integer indicating the number of classes to learn}

\item{allClassMeans}{A long data frame of methylation or bincount values with columns for position, methylation frequency (methFreq), class and replicate.}

\item{outFileBase}{A string that will be used in the filenames and titles of the plots produced}

\item{outPath}{path to directory where plots will be saved}

\item{xRange}{A vector of the first and last coordinates of the region to plot (default is c(-250,250))}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS (default="TSS")}

\item{baseFontSize}{The base font for the plotting theme (default=12 works well for 4x plots per A4 page)}

\item{figFormat}{format of output figures. Should be one of "png" or "pdf"}

\item{classesToPlot}{A numerical vector indicating which classes to plot (default NULL will plot all classes)}
}
\value{
None
}
\description{
Plot reads and class means with final classification
}
