% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmultigene.R
\name{plotGenesPerClass}
\alias{plotGenesPerClass}
\title{Plot number of genes per class}
\usage{
plotGenesPerClass(k_range, outPath, outFileBase)
}
\arguments{
\item{k_range}{range of number of classes}

\item{outPath}{Path for out put file}

\item{outFileBase}{the basename of the file of the matrices with classes
appended to the read name}
}
\value{
successful completion message
}
\description{
Plot number of genes per class
}
