% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{plotMatPCA}
\alias{plotMatPCA}
\title{Plot pca of data matrix}
\usage{
plotMatPCA(dataMatrix, classes)
}
\arguments{
\item{dataMatrix}{Matrix for PCA with row names containing __classNumber at
the end}

\item{classes}{Vactor of classifications for all the rows}
}
\value{
pca plots
}
\description{
Plot pca of data matrix
}
