% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{plotMatUMAP}
\alias{plotMatUMAP}
\title{Plot UMAP of data matrix}
\usage{
plotMatUMAP(dataMatrix, classes, rescale = T)
}
\arguments{
\item{dataMatrix}{Matrix for UMAP dimentionality reduction}

\item{classes}{Vector of classifications for all the rows}

\item{rescale}{Rescale matrix with 0-1 values to -1 to 1 values where -1 is
no methylation, NAs are 0 and 1 is methylation at that position.}
}
\value{
umap plots with different distance metrics
}
\description{
Plot UMAP of data matrix
}
