% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{plotPCAmatrices}
\alias{plotPCAmatrices}
\title{Plot matricies for different classifications}
\usage{
plotPCAmatrices(k_range, outPath, outFileBase)
}
\arguments{
\item{k_range}{range of number of classes}

\item{outPath}{Path for out put file}

\item{outFileBase}{the basename of the file of the matrices}
}
\value{
successful completion message
}
\description{
Plot matricies for different classifications
}
