% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{plotSmoothedClassMeans}
\alias{plotSmoothedClassMeans}
\title{Plot smoothed class means from multiple repeats}
\usage{
plotSmoothedClassMeans(
  allClassMeans,
  xRange = c(-250, 250),
  facet = FALSE,
  title = "Smoothed class means",
  myXlab = "CpG/GpC position",
  featureLabel = "TSS",
  baseFontSize = 12
)
}
\arguments{
\item{allClassMeans}{A long data frame of methylation or bincount values with columns for position, methylation frequency (methFreq), class and replicate.}

\item{xRange}{A vector of the first and last coordinates of the region to plot (default is c(-250,250))}

\item{facet}{Plot mean profiles separately as a facet_wrap plot (default=FALSE)}

\item{title}{A title for the plot (default is "Class means")}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS (default="TSS")}

\item{baseFontSize}{The base font for the plotting theme (default=12 works well for 4x plots per A4 page)}
}
\value{
Returns a ggplot2 object
}
\description{
Plot loess-smoothed class means from multiple repeats of EM.
}
