% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmultigene.R
\name{prepareWindows}
\alias{prepareWindows}
\title{Make window matrix}
\usage{
prepareWindows(fullMatrix, binSize = 20, stepSize = 1)
}
\arguments{
\item{fullMatrix}{A matrix containing methylation values for a full
region of the genome using coordinates relative to an anchor point, and
includes non-Cs that are marked by NaN. dataMatrix [i,j] is the methylation
value of sample i at position j.}

\item{binSize}{Size of window (in bp) in which to average results.}

\item{stepSize}{size of step by which to slide the window forward.}
}
\value{
Matrix with average fraction methylation within windows. Normalised
by the number of informative postions. NaN indicates a window with no
methylatable positions. NA indicates a window with no informative methylation
call.
}
\description{
Make window matrix
}
