% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{randomiseMatrixRows}
\alias{randomiseMatrixRows}
\title{Randomise the values in each row of a matrix}
\usage{
randomiseMatrixRows(dataMatrix)
}
\arguments{
\item{dataMatrix}{A matrix to be randomised}
}
\value{
A matrix where each row has been randomised
}
\description{
Randomise the values in each row of a matrix
}
