% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessMethMats.R
\name{recodeMatrixAsNumeric}
\alias{recodeMatrixAsNumeric}
\title{Recode matrix}
\usage{
recodeMatrixAsNumeric(dataMatrix)
}
\arguments{
\item{dataMatrix}{A matrix containing methylation values for a
region of the genome using coordinates relative to an anchor point.
dataMatrix[i,j] is the methylation value of sample i at position j.
0=not methylated, 1=methylated, NA=no methylation information at this
position.}
}
\value{
Matrix with methylation values coded on a -1 to 1 scale
}
\description{
Recode values in methylation matrix from a scale of 0 (no meth) to 1 (meth)
to a scale of -1 (no meth) to 1 (meth). NaN are set to 0 and NAs can
be either 0 or randomly assigned to -0.5 or 0.5
}
