% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessMethMats.R
\name{remove0}
\alias{remove0}
\title{Remove the number 0 from a vector}
\usage{
remove0(vec)
}
\arguments{
\item{vec}{A vector of numbers}
}
\value{
The same vector without the number 0
}
\description{
Remove the number 0 from a vector
}
\examples{
remove0(c(-2:2))
}
