% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessMethMats.R
\name{removeAllNArows}
\alias{removeAllNArows}
\title{Remove rows containing NAs}
\usage{
removeAllNArows(mat)
}
\arguments{
\item{mat}{A matrix of numbers and NAs.}
}
\value{
A matrix without the rows containing NAs
}
\description{
Remove rows containing NAs
}
\examples{
removeAllNArows(matrix(c(1,2,3,NA,4,5),nrow=3,byrow=T))
}
