% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifyByKnownClasses.R
\name{runClassLikelihoodRpts}
\alias{runClassLikelihoodRpts}
\title{Run several repeats classifying reads into classes}
\usage{
runClassLikelihoodRpts(
  dataMatrix,
  classes,
  numRepeats = 20,
  outPath = ".",
  xRange = c(-250, 250),
  outFileBase = "",
  myXlab = "CpG/GpC position",
  featureLabel = "TSS",
  baseFontSize = 12,
  distMetric = list(name = "euclidean")
)
}
\arguments{
\item{dataMatrix}{A matrix of methylation or bincount values (reads x position)}

\item{classes}{A matrix of classes x position giving the mean frequency of methylation at each position for each class}

\item{numRepeats}{An integer indicating the number of times to repeat the clustering (default=10)}

\item{outPath}{A string with the path to the directory where the output should go}

\item{xRange}{A vector of the first and last coordinates of the region to plot (default is c(-250,250))}

\item{outFileBase}{A string that will be used in the filenames and titles of the plots produced (default is "")}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS (default="TSS")}

\item{baseFontSize}{The base font for the plotting theme (default=12 works well for 4x plots per A4 page)}

\item{distMetric}{A list with the name of the distance metric and any
parameters it might require}
}
\value{
allClassMeans data.frame with columns: position, methFreq, class, replicate
}
\description{
Reads are classified into different classes based on likelihood estimation using the class means. Due to the need for randomisation when binarisation of the methylation matrix, the classification is carried out repeatedly and then the most frequently assigned class is chosen.
}
