% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{runEM}
\alias{runEM}
\title{Run EM iteratively to convergence}
\usage{
runEM(
  dataMatrix,
  numClasses,
  convergenceError = 1e-06,
  maxIterations = 100,
  printProgress = FALSE
)
}
\arguments{
\item{dataMatrix}{A matrix of methylation or bincount values (reads x position)}

\item{numClasses}{An integer indicating the number of classes to learn}

\item{convergenceError}{An float indicating the convergence threshold for stopping iteration}

\item{maxIterations}{An integer indicating the max number of iterations to perform even if the algorithm has not converged}

\item{printProgress}{Print messages showing progress of convergence}
}
\value{
list of three items:
            1) classes: a matrix with optimised classes (class x position)
            2) priorProb: a vector with the prior probabilities of each class
            3) posteriorProb: a matrix of probabilites of each sample belonging
            to a particular class (samples x class)
}
\description{
Run EM iteratively to convergence
}
