% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{runEMrepeats_withinSS}
\alias{runEMrepeats_withinSS}
\title{Run several repeats of iterative EM for randomised data}
\usage{
runEMrepeats_withinSS(
  dataMatrix,
  numClasses = 3,
  convergenceError = 1e-06,
  maxIterations = 100,
  EMrepeats = 10,
  distMetric = list(name = "euclidean")
)
}
\arguments{
\item{dataMatrix}{A matrix of methylation or bincount values (reads x position)}

\item{numClasses}{An integer indicating the number of classes to learn}

\item{convergenceError}{An float indicating the convergence threshold for stopping iteration}

\item{maxIterations}{An integer indicating the max number of iterations to perform even if the algorithm has not converged}

\item{EMrepeats}{An integer indicating the number of times to repeat the clustering (default=10)}

\item{distMetric}{A list with the name of the distance metric and any
parameters it might require}
}
\value{
Numeric. Total within cluster sum of squares.
}
\description{
Perform EM clustering on the same matrix several times. Necessary to check that classes are stable despite the random assignment of methylation fractions.
}
