% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMmultigene.R
\name{selectReadsFromMatrix}
\alias{selectReadsFromMatrix}
\title{Select reads from multiple genes}
\usage{
selectReadsFromMatrix(
  dataMatrix,
  minReads = 50,
  addToReadName = "",
  preferBest = T
)
}
\arguments{
\item{dataMatrix}{A matrix containing the sample. dataMatrix [i,j] is
the methylation value or bincount value of sample i at position j.}

\item{minReads}{The minimum number of reads a matrix needs to have to be
sampled from.}

\item{addToReadName}{A string that will be added to the read names (row names)
of the matrix}

\item{preferBest}{If TRUE will sample reads in inverse proportion to the
fraction of NAs in the read (takes more informative reads preferentially).
This is calculated using the fourth power of the fraction of NAs in order to
bias more strongly for better matrices.}
}
\value{
Returns a dataMatrix with minReads number of rows, subsampled from
the original dataMatrix.
}
\description{
Select reads from multiple genes
}
