% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{silhouettePlot}
\alias{silhouettePlot}
\title{Plot silhouette plot to evaluate classification}
\usage{
silhouettePlot(
  dataOrderedByClass,
  numClasses,
  outFileBase,
  EMrep = NULL,
  distMetric = list(name = "euclidean", rescale = F)
)
}
\arguments{
\item{dataOrderedByClass}{A matrix of methylation or bincount values (reads x position) that have been ordered by class. The assigned class, e.g. "__class1" etc has been appended to read names.}

\item{numClasses}{An integer with the number of classes learned}

\item{outFileBase}{A string that will be used in the filenames and titles of the plots produced}

\item{EMrep}{An integer indicating which EM repeat this is}

\item{distMetric}{A list with the name of the distance metric and any
parameters it might require}
}
\value{
A list of two items: a dataframe with silhouette stats and a silhouette plot object. The dataframe contains mean and SD of silhouette width overall, and per class, as well as number of reads per class
}
\description{
Create a silhouette plot to evaluate the quality of classification, and return some basic parameters about the classification
}
