% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{clusterRandomMatrices}
\alias{clusterRandomMatrices}
\title{Estimate within cluster sum of squares}
\usage{
clusterRandomMatrices(
  dataMatrix,
  k_range = 2:8,
  maxB = 100,
  convergenceError = 1e-06,
  maxIterations = 100,
  nThreads = 1,
  setSeed = F,
  distMetric = list(name = "euclidean")
)
}
\arguments{
\item{dataMatrix}{Matrix to be randomised and clustered}

\item{k_range}{A vector indicating different numbers of classes to learn}

\item{maxB}{The maximum number of randomisations to perform}

\item{convergenceError}{An float indicating the convergence threshold for stopping iteration}

\item{maxIterations}{An integer indicating the max number of iterations to perform even if the algorithm has not converged}

\item{nThreads}{Number of threads to use for generating background distribution (default is 1)}

\item{setSeed}{Logical value to determine if seed should be set for randomisation (default is FALSE)}

\item{distMetric}{A list with the name of the distance metric and any
parameters it might require}
}
\value{
A data frame with the average of the total within class sum of squares for multiple randomised matrices and different numbers of classes
}
\description{
Get estimated total within cluster sum of squares by clustering random matrices
}
