% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{getDistMatrix}
\alias{getDistMatrix}
\title{A generic function for implementing various distance metrics}
\usage{
getDistMatrix(
  binMat,
  distMetric = list(name = "euclidean", valNA = NULL, rescale = F)
)
}
\arguments{
\item{binMat}{A matrix of numbers for which you want to calculate the
distance between rows
#' @param valNA Value to give NAs in matrix}

\item{distMetric}{List continaing the name of the metric and the values of
any parameters the metric nromally takes.}
}
\value{
A distance object (lower triangle) with the distances between all
rows of the input matrix
}
\description{
To avoid writing too much code when i want to try a different distance
function i want a generic function that will be set up appropriately
with the right parameters (not specified before)
}
