% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{isMatrixValid}
\alias{isMatrixValid}
\title{Test if matrix is valid}
\usage{
isMatrixValid(
  dataMatrix,
  valueRange = c(0, 1),
  NAsValid = FALSE,
  checkColOrder = TRUE
)
}
\arguments{
\item{dataMatrix}{Matrix the values of which will be tested}

\item{valueRange}{Vector of min and max permissible values in the matrix
(default=c(0,1))}

\item{NAsValid}{Boolean TRUE/FALSE indicating if NAs are considered valid (default=FALSE)}

\item{checkColOrder}{Boolean TRUE/FALSE indicating if the order of column names should be checked. this assumes that these are numerical positions in ascending order (default=TRUE)}
}
\value{
Boolean TRUE/FALSE indicating if matrix contains only valid values
}
\description{
Test if matrix contains values that are only within the range specified by valueRange
}
