% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{plotMatUMAP}
\alias{plotMatUMAP}
\title{Plot UMAP of data matrix}
\usage{
plotMatUMAP(dataMatrix, classes)
}
\arguments{
\item{dataMatrix}{Matrix for UMAP dimentionality reduction}

\item{classes}{Vector of classifications for all the rows}
}
\value{
umap plots with different distance metrics
}
\description{
Plot UMAP of data matrix
}
