% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMbasic.R
\name{runEMrepeats}
\alias{runEMrepeats}
\title{Run several repeats of iterative EM}
\usage{
runEMrepeats(
  dataMatrix,
  numClasses = 3,
  convergenceError = 1e-06,
  maxIterations = 100,
  EMrepeats = 10,
  outPath = ".",
  xRange = c(-250, 250),
  outFileBase = "",
  myXlab = "CpG/GpC position",
  featureLabel = "TSS",
  baseFontSize = 12,
  doIndividualPlots = FALSE,
  distMetric = list(name = "euclidean", rescale = F)
)
}
\arguments{
\item{dataMatrix}{A matrix of methylation or bincount values (reads x position)}

\item{numClasses}{An integer indicating the number of classes to learn}

\item{convergenceError}{An float indicating the convergence threshold for stopping iteration}

\item{maxIterations}{An integer indicating the max number of iterations to perform even if the algorithm has not converged}

\item{EMrepeats}{An integer indicating the number of times to repeat the clustering (default=10)}

\item{outPath}{A string with the path to the directory where the output should go}

\item{xRange}{A vector of the first and last coordinates of the region to plot (default is c(-250,250))}

\item{outFileBase}{A string that will be used in the filenames and titles of the plots produced (default is "")}

\item{myXlab}{A label for the x axis (default is "CpG/GpC position")}

\item{featureLabel}{A label for a feature you want to plot, such as the position of the TSS (default="TSS")}

\item{baseFontSize}{The base font for the plotting theme (default=12 works well for 4x plots per A4 page)}

\item{doIndividualPlots}{Produce individual plots for each repeat (default=F)}

\item{distMetric}{A list with the name of the distance metric and any
parameters it might require}
}
\value{
allClassMeans data.frame with columns: position, methFreq, class, replicate
}
\description{
Perform EM clustering on the same matrix several times. Necessary to check ¨
that classes are stable despite the random assignment of methylation
fractions and the random assignment of starting conditions.
}
