% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{SumSqMatrixRows}
\alias{SumSqMatrixRows}
\title{Calculate the sum of square distances between rows of a matrix}
\usage{
SumSqMatrixRows(dataMatrix)
}
\arguments{
\item{dataMatrix}{A matrix for which to calcluate the sum of the sum of square distances between its rows.}
}
\value{
Numeric. The sum of the sum or square distances between the rows of the matrix.
}
\description{
Calculate the sum of square distances between rows of a matrix
}
