% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterUtils.R
\name{withinClusterSS}
\alias{withinClusterSS}
\title{Calculate within-cluster sum of squares}
\usage{
withinClusterSS(dataMatrix, classes)
}
\arguments{
\item{dataMatrix}{A matrix for which you wish to calculate the within cluster sum of squares distance between rows}

\item{classes}{A vector of classes designating which rows belong to which class}
}
\value{
A number indicating the sum of all within-cluster variation
}
\description{
Calculate within-cluster sum of squares
}
